+++ src/bunit.web/Extensions/InputFile/InputFileExtensions.cs
@@ -16,6 +16,16 @@ public static class InputFileExtensions
 	public static void UploadFiles(
 		this IRenderedComponent<InputFile> inputFileComponent,
 		params InputFileContent[] files)
+		=> UploadFilesAsync(inputFileComponent, files).GetAwaiter().GetResult();
+
+	/// <summary>
+	/// Uploads multiple files and invokes the OnChange event.
+	/// </summary>
+	/// <param name="inputFileComponent">The <see cref="InputFile"/> component which will upload the files.</param>
+	/// <param name="files">Files to upload.</param>
+	public static Task UploadFilesAsync(
+		this IRenderedComponent<InputFile> inputFileComponent,
+		params InputFileContent[] files)
 	{
 		if (inputFileComponent == null)
 			throw new ArgumentNullException(nameof(inputFileComponent));
@@ -23,6 +33,13 @@ public static class InputFileExtensions
 		if (!files.Any())
 			throw new ArgumentException("No files were provided to be uploaded.", nameof(files));
 
+		return UploadFilesInternal(inputFileComponent, files);
+	}
+
+	private static Task UploadFilesInternal(
+		IRenderedComponent<InputFile> inputFileComponent,
+		params InputFileContent[] files)
+	{
 		var browserFiles = files.Select(file => new BUnitBrowserFile(
 			file.Filename ?? string.Empty,
 			file.LastModified ?? default,
@@ -31,11 +48,7 @@ public static class InputFileExtensions
 			file.Content));
 
 		var args = new InputFileChangeEventArgs(browserFiles.ToArray());
-		var uploadTask = inputFileComponent.InvokeAsync(() => inputFileComponent.Instance.OnChange.InvokeAsync(args));
-		if (!uploadTask.IsCompleted)
-		{
-			uploadTask.GetAwaiter().GetResult();
-		}
+		return inputFileComponent.InvokeAsync(() => inputFileComponent.Instance.OnChange.InvokeAsync(args));
 	}
 }
 #endif