+++ tests/bunit.core.tests/TestContextBaseTest.net5.cs
@@ -2,6 +2,7 @@
 
 using Bunit.Rendering;
 using Bunit.TestAssets.SampleComponents.DisposeComponents;
+#pragma warning disable CA2252
 
 namespace Bunit;