+++ CHANGELOG.md
@@ -7,7 +7,10 @@ All notable changes to **bUnit** will be documented in this file. The project ad
 ## [Unreleased]
 
 ### Added
-- Added `DisposeComponentsAsync` as awaitable version of `DisposeComponents`. By [@linkdotnet](https://github.com/linkdotnet). 
+ - Added `DisposeComponentsAsync` as awaitable version of `DisposeComponents`. By [@linkdotnet](https://github.com/linkdotnet). 
+
+### Fixed
+ - The created HTML contained encoded strings. Reported by [@tobiasbrandstaedter](https://github.com/tobiasbrandstaedter). Fixed by [@linkdotnet](https://github.com/linkdotnet).
 
 ## [1.11.7] - 2022-10-13