+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to **bUnit** will be documented in this file. The project ad
 ### Fixed
  - The created HTML contained encoded strings. Reported by [@tobiasbrandstaedter](https://github.com/tobiasbrandstaedter). Fixed by [@linkdotnet](https://github.com/linkdotnet).
 
+### Added
+- Added `UploadFilesAsync` as awaitable version of `UploadFiles`. By [@linkdotnet](https://github.com/linkdotnet).
+
 ## [1.11.7] - 2022-10-13
 
 ### Added